requirements.txt
astroid==2.4.2
certifi==2020.6.20
Django==2.2
django-ckeditor==5.9.0
django-crispy-forms==1.9.2
django-js-asset==1.2.2
gunicorn==20.0.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
Pillow==7.2.0
pylint==2.5.3
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
whitenoise==5.2.0
wrapt==1.12.1
